reuse the same container for edit and upload page
